Release 5.10.2.2 FIX: Hardware token import was not always working correctly for 013 appliances FIX: PHP 8.x incompatibility issue in SMTP library (Thanks to @dexkuz) ENH: PHP 8.x deprecated code cleaned ENH: Yubikey import now also accept "Yubico CSV" format ENH: New 2FA opt-in option, to migrate smoothly from Without2FA accounts to 2FA ENH: Better rights handling in the Docker version ENH: Embedded Windows PHP edition updated to version 8.5.4 x64 ENH: Embedded Windows nginx edition updated to version 1.29.7 ENH: Documentation enhancement

+++ contrib/MultiotpShowStatus.php
@@ -66,7 +66,7 @@ function MultiotpShowStatus($done, $total, $size=30)
     if(empty($start_time)) $start_time=time();
     $now = time();
 
-    $perc=(double)($done/$total);
+    $perc=(float)($done/$total);
 
     $bar=floor($perc*$size);
